First Commit
This will be the Front End of "Der MarktPlatz", primarily based on the Angular Framework. Following things have been added to the structure, which will function based on the "Mein MarktPlatz" Project:

A Home Page
A Search Bar
An Item Template

More such components (and services) will be added soon.